# ComponentsProviderGenerator - lower group limit for addRemovedBeans()