- to eliminate the perf regression caused by
MethodWriter#computeAllFrames() and Frame#merge()
- resolves #24712