@Postremus 2022-04-04
average of 3 runs:
2.8.0.CR1
mvn quarkus:dev 2304ms
force restart 578ms

2.7.5.Final
mvn quarkus:dev 1935ms
force restart 235ms

This PR
mvn quarkus:dev 1982ms
force restart 273ms

main (9aab340b) with cf388350609ff4b108dacd09a187523d35b6fc4c reverted
mvn quarkus:dev 1935ms
force restart 230ms

2.7.5.Final still seems to be about 40ms faster compared to this pr.
Reverting the mentioned commit on main results in about the same times.

I am personally ok with merging this one as a first step. Arc is simply doing more stuff, which might be an execuse for taking a little bit longer.