@geoand 2022-04-05
What does the group limit change do in practice?