@mkouba 2022-04-05
> What does the group limit change do in practice?

The generated `io.quarkus.arc.ComponentsProvider` has several `addRemovedBeansX()` methods that instantiate the runtime metadata about removed beans (`io.quarkus.arc.RemovedBean`). The group limit specifies how many components are "grouped" in a single `addRemovedBeansX()` method. In other words, lowering the limit means _more_ `addRemovedBeansX()` methods with _shorter method body_.